refactor: organize app routes
