+++ src/app/App.stories.tsx
@@ -7,10 +7,10 @@
 import type { Meta, StoryObj } from "@storybook/react";
 import { expect } from "storybook/test";
 
+import { AppRoutes } from "@/app/routes";
 import { type PageStoryParameters, preparePageStory, withPageStory } from "@/storybook/PageDecorator";
 import { PAGE_STORY_CARD_ID, PAGE_STORY_DECK_ID, pageStoryState } from "@/storybook/pageFixture";
 import { STORYBOOK_DECK_IMPORT_URL } from "@/storybook/handlers";
-import { AppRoutes } from "./App";
 
 const page = (path: string, overrides: Partial<Omit<PageStoryParameters, "path">> = {}): PageStoryParameters => ({
   ...pageStoryState,
