+++ src/app/routes/index.ts
@@ -1,0 +1,2 @@
+export { AppRoutes } from "./AppRoutes";
+export type { Page, PageKey } from "./types";
